Aplicado Black
