+++ comedores/models/comedor.py
@@ -304,6 +304,3 @@ class ValorComida(models.Model):
     tipo = models.CharField(max_length=50)
     valor = models.DecimalField(max_digits=10, decimal_places=2)
     fecha = models.DateField()
-
-
-
