+++ comedores/urls/web_urls.py
@@ -23,12 +23,6 @@ from comedores.views.web_views import (
     NominaCreateView,
     NominaDeleteView,
     NominaUpdateView,
-    AdmisionesTecnicosListView,
-    AdmisionesTecnicosCreateView,
-    AdmisionesTecnicosUpdateView,
-    sub_estados_intervenciones_ajax,
-    subir_archivo_admision,
-    eliminar_archivo_admision,
 )
 from configuraciones.decorators import group_required
 
@@ -142,35 +136,5 @@ urlpatterns = [
         "comedores/nomina/borrar/<pk>/<pk2>",
         group_required("Comedores")(NominaDeleteView.as_view()),
         name="nomina_borrar",
-    ),
-    path(
-        "comedores/ajax/load-subestadosintervenciones/",
-        sub_estados_intervenciones_ajax,
-        name="ajax_load_subestadosintervenciones",
-    ),
-    path(
-        "comedores/admisiones/tecnicos/listar",
-        group_required("Comedores")(AdmisionesTecnicosListView.as_view()),
-        name="admisiones_tecnicos_listar",
-    ),
-    path(
-        "comedores/admisiones/tecnicos/crear/<pk>",
-        group_required("Comedores")(AdmisionesTecnicosCreateView.as_view()),
-        name="admisiones_tecnicos_crear",
-    ),
-    path(
-        "comedores/admisiones/tecnicos/editar/<pk>",
-        group_required("Comedores")(AdmisionesTecnicosUpdateView.as_view()),
-        name="admisiones_tecnicos_editar",
-    ),
-    path(
-        "admision/<int:admision_id>/documentacion/<int:documentacion_id>/subir/",
-        subir_archivo_admision,
-        name="subir_archivo_admision",
-    ),
-    path(
-        "admision/<int:admision_id>/documentacion/<int:documentacion_id>/eliminar/",
-        eliminar_archivo_admision,
-        name="eliminar_archivo_admision",
     ),
 ]
